requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
future==0.18.2
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-dialogflow==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
packaging==20.9
proto-plus==1.18.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
